Improve network issue by duplicating IP address.
If exist network has same subnet as demo network,
network hangs up and all communication is stopped.
To avoid this issue, new subnet which seems not using much is assigned
by this patch.
However, network configulation depends on each environment
so that this patch can not avoid this issue completely

Signed-off-by: Yuya Hamamachi <yuya.hamamachi.sx@renesas.com>