cr52: Change directory to store venv envrionment
Currently, .venv directory is stored in working directoy.
So, it removes when clean build flag is enabled.
However, venv is build envrionment as same as Zephyr SDK,
so it's not needed to setup again.
This patch moves to .venv into Zephyr SDK directory which is not removed
when clean build flag is enabled.

Signed-off-by: Yuya Hamamachi <yuya.hamamachi.sx@renesas.com>