tool: IPL: Add burn_s4.py
Currently, there is gap in how to flash loader between Linux and
Windows PC.
This is due to burn_s4.sh can be only run on Unix shell environment.
Thus, we have prepared Teraterm Macro for Windows envrionment.

This patch adds new script written in python.
python can be used both Host PC envrionment so that we can remove the
gap in how to flash loader.

Signed-off-by: Yuya Hamamachi <yuya.hamamachi.sx@renesas.com>