g4mh: Add warning message about CC-RH compiler for Linux
CC-RH compiler for Windows with wine was used so far.
But, compiler for Linux has been used in latest Whitebox SDK.
This patch notifies users to switch the compiler.

Note:
As of now, compiler for Windows with Wine envrionment can built
each software, but we will not continue to support it.
Because Wine was temporary solution until compiler for Linux was released.

Signed-off-by: Yuya Hamamachi <yuya.hamamachi.sx@renesas.com>