ca55: rcar-common: grafana: Fix build error
Release-9.2.1 branch is removed in upstream and this commit doesn't
belongs to any branch as of now.
Thus, this patch changes to not specify the branch using nobranch.
And, latest wire package is not compatible go v1.18 so that this patch
chagnes to use v0.6.0 instead of latest.

This patch is based on following conversation:
https://community.renesas.com/automotive-soc/gateway/f/forum/56279/r-car-s4-whitebox-sdk-build-error

Signed-off-by: Yuya Hamamachi <yuya.hamamachi.sx@renesas.com>